Entities
Entities

+++ SchoolPEPProject/src/main/java/com/nimblix/SchoolPEPProject/Model/Student.java
@@ -13,7 +13,7 @@ import lombok.*;
 public class Student {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(name = "full_name")
     private String fullName;
@@ -25,7 +25,7 @@ public class Student {
     private String password;
 
     @Column(name = "schoolId")
-    private Integer schoolId;
+    private Long schoolId;
 
     @Column(name = "status")
     private String status;
