+++ SchoolPEPProject/src/main/java/com/nimblix/SchoolPEPProject/Request/StudentRegistrationRequest.java
@@ -12,6 +12,6 @@ public class StudentRegistrationRequest {
     private String email;
     private String password;
     private String reEnterPassword;
-    private Integer schoolId;
+    private Long schoolId;
 
 }
